Rename plist_ipv4.rs to prefix_set.rs.
