+++ zebra-rs/yang/config.yang
@@ -587,7 +587,10 @@ module config {
       list "prefix" {
         key "prefix";
         leaf "prefix" {
-          type inet:ipv4-prefix;
+          type union {
+            type inet:ipv4-prefix;
+            type inet:ipv6-prefix;
+          }
         }
         leaf le {
           type uint8;
